flytter mottaker henting til egen tjeneste
