+++ formidling/src/main/java/no/nav/ung/sak/formidling/GenerertBrev.java
@@ -2,6 +2,7 @@ package no.nav.ung.sak.formidling;
 
 import no.nav.ung.kodeverk.dokument.DokumentMalType;
 import no.nav.ung.kodeverk.formidling.TemplateType;
+import no.nav.ung.sak.formidling.mottaker.PdlPerson;
 import no.nav.ung.sak.formidling.pdfgen.PdfGenDokument;
 
 public record GenerertBrev(
