+++ formidling/src/main/java/no/nav/ung/sak/formidling/VedtaksbrevGenerererTjenesteImpl.java
@@ -4,27 +4,20 @@ package no.nav.ung.sak.formidling;
 import io.opentelemetry.instrumentation.annotations.WithSpan;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import no.nav.ung.sak.behandlingslager.behandling.Behandling;
-import no.nav.ung.sak.behandlingslager.behandling.personopplysning.PersonopplysningEntitet;
-import no.nav.ung.sak.behandlingslager.behandling.personopplysning.PersonopplysningGrunnlagEntitet;
-import no.nav.ung.sak.behandlingslager.behandling.personopplysning.PersonopplysningRepository;
 import no.nav.ung.sak.behandlingslager.behandling.repository.BehandlingRepository;
 import no.nav.ung.sak.behandlingslager.formidling.VedtaksbrevValgEntitet;
 import no.nav.ung.sak.behandlingslager.formidling.VedtaksbrevValgRepository;
-import no.nav.ung.sak.domene.person.pdl.AktørTjeneste;
 import no.nav.ung.sak.formidling.innhold.ManuellVedtaksbrevInnholdBygger;
 import no.nav.ung.sak.formidling.innhold.VedtaksbrevInnholdBygger;
+import no.nav.ung.sak.formidling.mottaker.BrevMottakerTjeneste;
 import no.nav.ung.sak.formidling.pdfgen.PdfGenDokument;
 import no.nav.ung.sak.formidling.pdfgen.PdfGenKlient;
 import no.nav.ung.sak.formidling.template.TemplateInput;
 import no.nav.ung.sak.formidling.template.dto.TemplateDto;
 import no.nav.ung.sak.formidling.template.dto.felles.FellesDto;
 import no.nav.ung.sak.formidling.template.dto.felles.MottakerDto;
-import no.nav.ung.sak.typer.AktørId;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.Objects;
 
 @ApplicationScoped
 public class VedtaksbrevGenerererTjenesteImpl implements VedtaksbrevGenerererTjeneste {
@@ -32,31 +25,27 @@ public class VedtaksbrevGenerererTjenesteImpl implements VedtaksbrevGenerererTje
     private static final Logger LOG = LoggerFactory.getLogger(VedtaksbrevGenerererTjenesteImpl.class);
 
     private BehandlingRepository behandlingRepository;
-    private AktørTjeneste aktørTjeneste;
     private PdfGenKlient pdfGen;
-    private PersonopplysningRepository personopplysningRepository;
     private VedtaksbrevValgRepository vedtaksbrevValgRepository;
     private ManuellVedtaksbrevInnholdBygger manuellVedtaksbrevInnholdBygger;
+    private BrevMottakerTjeneste brevMottakerTjeneste;
 
     private VedtaksbrevRegler vedtaksbrevRegler;
 
     @Inject
     public VedtaksbrevGenerererTjenesteImpl(
         BehandlingRepository behandlingRepository,
-        AktørTjeneste aktørTjeneste,
         PdfGenKlient pdfGen,
-        PersonopplysningRepository personopplysningRepository,
         VedtaksbrevRegler vedtaksbrevRegler,
         VedtaksbrevValgRepository vedtaksbrevValgRepository,
-        ManuellVedtaksbrevInnholdBygger manuellVedtaksbrevInnholdBygger) {
+        ManuellVedtaksbrevInnholdBygger manuellVedtaksbrevInnholdBygger, BrevMottakerTjeneste brevMottakerTjeneste) {
 
         this.behandlingRepository = behandlingRepository;
-        this.aktørTjeneste = aktørTjeneste;
         this.pdfGen = pdfGen;
-        this.personopplysningRepository = personopplysningRepository;
         this.vedtaksbrevRegler = vedtaksbrevRegler;
         this.vedtaksbrevValgRepository = vedtaksbrevValgRepository;
         this.manuellVedtaksbrevInnholdBygger = manuellVedtaksbrevInnholdBygger;
+        this.brevMottakerTjeneste = brevMottakerTjeneste;
     }
 
     public VedtaksbrevGenerererTjenesteImpl() {
@@ -112,7 +101,7 @@ public class VedtaksbrevGenerererTjenesteImpl implements VedtaksbrevGenerererTje
 
         VedtaksbrevInnholdBygger bygger = regelResultat.automatiskVedtaksbrevBygger();
         var resultat = bygger.bygg(behandling, regelResultat.detaljertResultatTimeline());
-        var pdlMottaker = hentMottaker(behandling);
+        var pdlMottaker = brevMottakerTjeneste.hentMottaker(behandling);
         var input = new TemplateInput(resultat.templateType(),
             new TemplateDto(
                 FellesDto.automatisk(new MottakerDto(pdlMottaker.navn(), pdlMottaker.fnr())),
@@ -143,7 +132,7 @@ public class VedtaksbrevGenerererTjenesteImpl implements VedtaksbrevGenerererTje
     private GenerertBrev doGenererManuellVedtaksbrev(Long behandlingId, boolean kunHtml) {
         var behandling = behandlingRepository.hentBehandling(behandlingId);
         var resultat = manuellVedtaksbrevInnholdBygger.bygg(behandling, null);
-        var pdlMottaker = hentMottaker(behandling);
+        var pdlMottaker = brevMottakerTjeneste.hentMottaker(behandling);
         var input = new TemplateInput(resultat.templateType(),
             new TemplateDto(
                 FellesDto.manuell(new MottakerDto(pdlMottaker.navn(), pdlMottaker.fnr())),
@@ -159,22 +148,6 @@ public class VedtaksbrevGenerererTjenesteImpl implements VedtaksbrevGenerererTje
             resultat.dokumentMalType(),
             resultat.templateType()
         );
-    }
-
-    private PdlPerson hentMottaker(Behandling behandling) {
-        PersonopplysningGrunnlagEntitet personopplysningGrunnlagEntitet = personopplysningRepository.hentPersonopplysninger(behandling.getId());
-        PersonopplysningEntitet personopplysning = personopplysningGrunnlagEntitet.getGjeldendeVersjon().getPersonopplysning(behandling.getAktørId());
-
-        String navn = personopplysning.getNavn();
-
-        AktørId aktørId = behandling.getFagsak().getAktørId();
-        var personIdent = aktørTjeneste.hentPersonIdentForAktørId(aktørId)
-            .orElseThrow(() -> new IllegalArgumentException("Fant ikke person med aktørid"));
-
-        String fnr = personIdent.getIdent();
-        Objects.requireNonNull(fnr);
-
-        return new PdlPerson(fnr, aktørId, navn);
     }
