+++ formidling/src/test/java/no/nav/ung/sak/formidling/VedtaksbrevGenerererTjenesteFake.java
@@ -2,6 +2,7 @@ package no.nav.ung.sak.formidling;
 
 import no.nav.ung.kodeverk.dokument.DokumentMalType;
 import no.nav.ung.kodeverk.formidling.TemplateType;
+import no.nav.ung.sak.formidling.mottaker.PdlPerson;
 import no.nav.ung.sak.formidling.pdfgen.PdfGenDokument;
 import no.nav.ung.sak.typer.AktørId;
